+++ src/app/log-in/log-in.component.css
@@ -1,0 +1,26 @@
+
+.buttonRow button{
+  width: 100px;
+  margin-right:8px;
+}
+
+mat-card{
+  max-width: 300px;
+  text-align: center;
+}
+
+.main-div {
+  position: absolute;
+  left: 50%;
+  top: 50%;
+  transform: translate(-50%, -50%); 
+}
+
+
+.container {
+  height: 100%;
+  display: flex;
+
+  background: url(../../assets/background.jpg) no-repeat center fixed;
+  -webkit-background-size: cover; /* pour anciens Chrome et Safari */
+  background-size: cover; /* version standardisée */}
