+++ src/app/open-case.service.ts
@@ -1,6 +1,6 @@
 import { FormBuilder, Validators, FormControl, FormGroup } from '@angular/forms';
 import { Injectable } from '@angular/core';
-import { ICaseFields } from './open-case/caseFields';
+import { ICaseFields, CaseTypes } from './open-case/caseFields';
 
 @Injectable({
   providedIn: 'root'
@@ -13,66 +13,12 @@ export class OpenCaseService {
   }
   )
 
-
-  CaseTypes = {
-    "Accident":[
-      {
-        fieldDisplay: "Name",
-        priority: 1,
-        fieldIdentifier: "clientName",
-        fieldType: "string"
-      },
-      {
-        fieldDisplay: "Identity Card #",
-        priority: 2,
-        fieldIdentifier: "accidentDate",
-        fieldType: "string"
-      },
-      {
-        fieldDisplay: "License Plate #",
-        priority: 4,
-        fieldIdentifier: "licensePlateNumber",
-        fieldType: "string"
-      },
-      {
-        fieldDisplay: "Accident Date",
-        priority: 3,
-        fieldIdentifier: "accidentDate",
-        fieldType: "date"
-      }],
-    
-    "Other": [
-      {
-        fieldDisplay: "Dummy 1",
-        priority: 1,
-        fieldIdentifier: "dummy1",
-        fieldType: "string"
-      },
-      {
-        fieldDisplay: "Dummy 2",
-        priority: 2,
-        fieldIdentifier: "dummy2",
-        fieldType: "date"
-      },
-      {
-        fieldDisplay: "Dummy 3",
-        priority: 3,
-        fieldIdentifier: "dummy3",
-        fieldType: "string"
-      }
-    ]
-  }
-
-
-
-
-  
   getFormItems(){
     return this.caseForm
   }
 
   getCaseTypes(){
-    return this.CaseTypes
+    return CaseTypes
   }
 
   checkValues() {
@@ -88,7 +34,6 @@ export class OpenCaseService {
       }
     }
   }
-  
   
   sendFormItems(input:FormGroup){
     this.caseForm = input
