--- src/app/open-case/open-case-confirm/open-case-confirm.component.html
@@ -1,3 +1,0 @@
-<p>open-case-confirm works!</p>
-
-<router-outlet></router-outlet>
