--- src/app/open-case/open-case-confirm/open-case-confirm.component.ts
@@ -1,15 +1,0 @@
-import { Component, OnInit } from '@angular/core';
-
-@Component({
-  selector: 'app-open-case-confirm',
-  templateUrl: './open-case-confirm.component.html',
-  styleUrls: ['./open-case-confirm.component.scss']
-})
-export class OpenCaseConfirmComponent implements OnInit {
-
-  constructor() { }
-
-  ngOnInit() {
-  }
-
-}
