+++ src/app/open-case/open-case-routing.module.ts
@@ -1,8 +1,7 @@
-import { OpenCaseConfirmComponent } from './open-case-confirm/open-case-confirm.component';
 import { CommonModule } from '@angular/common';
 import { NgModule } from '@angular/core';
 import { Routes, RouterModule } from '@angular/router';
-import { OpenCaseComponent } from './open-case/open-case.component';
+import { OpenCaseComponent } from './open-case.component';
 
 
 const routes: Routes = [
@@ -17,12 +16,7 @@ const routes: Routes = [
     //     component: OpenCaseConfirmComponent
     //   }
     // ]
-  },
-  {
-    path:'opencaseconfirm',
-    component: OpenCaseConfirmComponent
   }
-  
 ];
 
 @NgModule({
