+++ src/app/open-case/open-case.component.scss
@@ -1,12 +1,3 @@
-/* .example-container {
-    display: flex;
-    flex-direction: column;
-  }
-  
-  .example-container > * {
-    width: 100%;
-  } */
-
 .buttonRow button{
   width: 100px;
   margin-right:8px;
@@ -18,10 +9,6 @@ mat-card{
 }
 
 .main-div {
-  /* display: flex;
-  justify-content: center;
-  align-items: center; */
-
   position: absolute;
   left: 50%;
   top: 50%;
@@ -33,8 +20,6 @@ mat-card{
 .container {
   height: 100%;
   display: flex;
-  /* align-items: center; */
-  /* justify-content: center; */
 
   background: url(../../assets/background.jpg) no-repeat center fixed;
   -webkit-background-size: cover; /* pour anciens Chrome et Safari */
