--- src/app/open-case/open-case/open-case.component.scss
@@ -1,42 +1,0 @@
-/* .example-container {
-    display: flex;
-    flex-direction: column;
-  }
-  
-  .example-container > * {
-    width: 100%;
-  } */
-
-  .buttonRow button{
-    width: 100px;
-    margin-right:8px;
-  }
-  
-  mat-card{
-    max-width: 300px;
-    text-align: center;
-  }
-  
-  .main-div {
-    /* display: flex;
-    justify-content: center;
-    align-items: center; */
-  
-    position: absolute;
-    left: 50%;
-    top: 50%;
-    transform: translate(-50%, -50%);
-    
-  }
-  
-  
-  .container {
-    height: 100%;
-    display: flex;
-    /* align-items: center; */
-    /* justify-content: center; */
-  
-    background: url(../../../assets/background.jpg) no-repeat center fixed;
-    -webkit-background-size: cover; /* pour anciens Chrome et Safari */
-    background-size: cover; /* version standardisée */}
-
